rename egg to qt-smoke
Rationale: although the library that we are binding is called
"libsmokeqt", the relevant thing for this egg is that these are Qt
bindings.  The fact that smoke is used as middleman is a secondary point,
and had the egg name "qt" not already been taken, it would not even bear
mentioning so prominently.

This patch just renames the egg, not the component modules.  Will do that
in the next patch.